Update .cbp #10
Also counting

+++ source/Libirid_Server.h
@@ -22,7 +22,7 @@ class Libirid_Server
         void finalize();
 
         //
-        int numbah;
+        long long numbah;
 
         //
         void run();
